Build : seperate parser header & move insert_* to insert_data/
