+++ .gitignore
@@ -2,8 +2,7 @@
 # (1)program
 minishell
 include/*
-src/manda/include/*
-src/bonus/include/*
+
 # (2)etc.
 test*
 .DS_Store
