+++ src/manda/Makefile
@@ -9,7 +9,8 @@ NAME := minishell
 SRCS := main_syntax_analyzer_print.c \
 
 # SRCS(in SRCDIR) ------
-DIRS := parser
+DIRS := insert_data \
+		parser \
 
 SUBS := $(addsuffix .a,$(DIRS))
