+++ src/manda/insert_data/insert_lr_table.h
