+++ src/manda/insert_data/insert_util.c
