+++ src/manda/parser/tokenizer.h
@@ -1,21 +1,21 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   parser.h                                           :+:      :+:    :+:   */
+/*   tokenizer.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sehwjang <sehwjang@student.42.fr>          +#+  +:+       +#+        */
+/*   By: taerakim <taerakim@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/22 14:01:48 by taerakim          #+#    #+#             */
-/*   Updated: 2024/04/05 13:27:11 by taerakim         ###   ########.fr       */
+/*   Updated: 2024/04/05 16:40:37 by taerakim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef PARSER_H
-# define PARSER_H
+#ifndef TOKENIZER_H
+# define TOKENIZER_H
 
-# include <stdbool.h>
 # include <stdlib.h>
 # include "libft.h"
+# include "input_type.h"
 
 /* ******************************** LITERAL ********************************* */
 # define S_PIPE "|"
@@ -28,66 +28,12 @@
 # define S_LPAREN "("
 # define S_RPAREN ")"
 
-/* ********************************** TYPE ********************************** */
-typedef enum e_kind
-{
-	none = -1,
-	terminal,
-	non_terminal,
-	state
-}	t_kind;
-
-typedef enum e_terminal
-{
-	and_if = 0,
-	or_if,
-	pipe,
-	lparen,
-	rparen,
-	word,
-	less,
-	great,
-	dgreat,
-	dless,
-	dollar_sign,
-	undefined
-}	t_termi;
-
-typedef enum e_non_terminal
-{
-	complete_command = 0,
-	and_or,
-	pipeline,
-	command,
-	subshell,
-	simple_command,
-	cmd_word,
-	cmd_suffix,
-	redirect_list,
-	io_redirect,
-	io_file,
-	filename,
-	io_here,
-	here_end
-}	t_ntermi;
-
 /* ********************************* STRUCT ********************************* */
-# define LEFT 0
-# define MID 1
-# define RIGHT 2
-
 typedef struct s_token
 {
 	t_termi	type;
 	char	*word;
 }			t_token;
-
-typedef struct s_parse_tree
-{
-	t_ntermi	type;
-	t_kind		child_type[3];
-	void		*child[3];
-}				t_parse_tree;
 
 /* ******************************** FUNCTION ******************************** */
 t_list	*tokenizer(char *command);
